Small rename to long term memory (#6914)
